# docs/swagger.yaml
basePath: /
definitions:
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  entities.AlertaMantenimiento:
    properties:
      alerta_id:
        type: integer
      atendido:
        type: boolean
      camion_id:
        type: integer
      created_at:
        type: string
      descripcion:
        type: string
      observaciones:
        type: string
      tipo_mantenimiento_id:
        type: integer
    type: object
...
  entities.AlertaMantenimientoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.AlertaMantenimiento'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.CreateRegistroMantenimientoRequest:
    properties:
      alerta_id:
        type: integer
      camion_id:
        type: integer
      coordinador_id:
        type: integer
      fecha_realizada:
        type: string
      kilometraje_mantenimiento:
        type: number
      mecanico_responsable:
        type: string
      observaciones:
        type: string
    required:
    - camion_id
    type: object
...
  entities.RegistroMantenimiento:
    properties:
      alerta_id:
        type: integer
      camion_id:
        type: integer
      coordinador_id:
        type: integer
      created_at:
        type: string
      fecha_realizada:
        type: string
      kilometraje_mantenimiento:
        type: number
      mecanico_responsable:
        type: string
      observaciones:
        type: string
      registro_id:
        type: integer
    type: object
  entities.RegistroMantenimientoListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/entities.RegistroMantenimiento'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.RegistroMantenimientoMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.RegistroMantenimientoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/entities.RegistroMantenimiento'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  entities.UpdateRegistroMantenimientoRequest:
    properties:
      alerta_id:
        type: integer
      camion_id:
        type: integer
      coordinador_id:
        type: integer
      fecha_realizada:
        type: string
      kilometraje_mantenimiento:
        type: number
      mecanico_responsable:
        type: string
      observaciones:
        type: string
    type: object
...
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/alertas-mantenimiento/camion/{camion_id}:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AlertaMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Alertas por camión
      tags:
      - AlertaMantenimiento
  /api/alertas-mantenimiento/fecha:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AlertaMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Alertas por rango de fecha
      tags:
      - AlertaMantenimiento
...
  /api/alertas-mantenimiento/tipo/{tipo_id}:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AlertaMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Alertas por tipo de mantenimiento
      tags:
      - AlertaMantenimiento
...
  /api/registros-mantenimiento/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar registros de mantenimiento
      tags:
      - RegistroMantenimiento
    post:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CreateRegistroMantenimientoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear registro de mantenimiento
      tags:
      - RegistroMantenimiento
...
  /api/registros-mantenimiento/alerta/{alerta_id}:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registros por alerta
      tags:
      - RegistroMantenimiento
...
  /api/registros-mantenimiento/coordinador/{coordinador_id}:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registros por coordinador
      tags:
      - RegistroMantenimiento
...
  /api/reportes-mantenimiento-generado/fecha:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reportes por rango de fecha
      tags:
      - ReporteMantenimientoGenerado
...
  /api/tipos-mantenimiento/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar tipo de mantenimiento
      tags:
      - TipoMantenimiento
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tipo de mantenimiento por ID
      tags:
      - TipoMantenimiento
    put:
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateRegistroMantenimientoRequest'
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RegistroMantenimientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Actualizar tipo de mantenimiento
      tags:
      - TipoMantenimiento
securityDefinitions:
  BearerAuth:
    description: 'Escribe ''Bearer '' seguido de tu token JWT. Ejemplo: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"